Add standard test
