+++ spring-boot-integration/h2-demo-app/src/test/java/io/github/mfvanek/pg/spring/h2/H2DemoApplicationTest.java
@@ -20,8 +20,8 @@ import javax.sql.DataSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@ActiveProfiles("test")
 @SpringBootTest
+@ActiveProfiles("test")
 class H2DemoApplicationTest {
 
     @Autowired
