+++ spring-boot-integration/kotlin-demo-app/src/test/kotlin/io/github/mfvanek/pg/spring/postgres/kt/DatabaseStructureStaticAnalysisTest.kt
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) 2019-2024. Ivan Vakhrushev and others.
+ * https://github.com/mfvanek/pg-index-health
+ *
+ * This file is a part of "pg-index-health" - a Java library for
+ * analyzing and maintaining indexes health in PostgreSQL databases.
+ *
+ * Licensed under the Apache License 2.0
+ */
+
+package io.github.mfvanek.pg.spring.postgres.kt
+
+import io.github.mfvanek.pg.core.checks.common.DatabaseCheckOnHost
+import io.github.mfvanek.pg.core.checks.common.Diagnostic
+import io.github.mfvanek.pg.model.dbobject.DbObject
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.Test
+import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.test.context.ActiveProfiles
+
+@SpringBootTest
+@ActiveProfiles("test")
+internal class DatabaseStructureStaticAnalysisTest {
+
+    @Autowired
+    private lateinit var checks: List<DatabaseCheckOnHost<out DbObject?>>
+
+    @Test
+    fun checksShouldWork() {
+        assertThat(checks)
+            .hasSameSizeAs(Diagnostic.entries.toTypedArray())
+
+        checks
+            .filter { it.isStatic }
+            .forEach {
+                assertThat(it.check())
+                    .`as`(it.diagnostic.name)
+                    .isEmpty()
+            }
+    }
+}
