+++ spring-boot-integration/postgres-demo-app/src/test/java/io/github/mfvanek/pg/spring/postgres/DatabaseStructureStaticAnalysisTest.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright (c) 2019-2024. Ivan Vakhrushev and others.
+ * https://github.com/mfvanek/pg-index-health
+ *
+ * This file is a part of "pg-index-health" - a Java library for
+ * analyzing and maintaining indexes health in PostgreSQL databases.
+ *
+ * Licensed under the Apache License 2.0
+ */
+
+package io.github.mfvanek.pg.spring.postgres;
+
+import io.github.mfvanek.pg.core.checks.common.DatabaseCheckOnHost;
+import io.github.mfvanek.pg.core.checks.common.Diagnostic;
+import io.github.mfvanek.pg.model.dbobject.DbObject;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@SpringBootTest
+@ActiveProfiles("test")
+class DatabaseStructureStaticAnalysisTest {
+
+    @Autowired
+    private List<DatabaseCheckOnHost<? extends DbObject>> checks;
+
+    @Test
+    void checksShouldWork() {
+        assertThat(checks)
+            .hasSameSizeAs(Diagnostic.values());
+
+        checks.stream()
+            .filter(DatabaseCheckOnHost::isStatic)
+            .forEach(c ->
+                assertThat(c.check())
+                    .as(c.getDiagnostic().name())
+                    .isEmpty());
+    }
+}
