+++ spring-boot-integration/postgres-demo-app/src/test/java/io/github/mfvanek/pg/spring/postgres/PostgresDemoApplicationTest.java
@@ -12,9 +12,6 @@ package io.github.mfvanek.pg.spring.postgres;
 
 import com.zaxxer.hikari.HikariDataSource;
 import io.github.mfvanek.pg.connection.PgConnection;
-import io.github.mfvanek.pg.core.checks.common.DatabaseCheckOnHost;
-import io.github.mfvanek.pg.core.checks.common.Diagnostic;
-import io.github.mfvanek.pg.model.dbobject.DbObject;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -22,12 +19,10 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.core.env.Environment;
 import org.springframework.test.context.ActiveProfiles;
 
-import java.util.List;
-
 import static org.assertj.core.api.Assertions.assertThat;
 
-@ActiveProfiles("test")
 @SpringBootTest
+@ActiveProfiles("test")
 class PostgresDemoApplicationTest {
 
     @Autowired
@@ -35,9 +30,6 @@ class PostgresDemoApplicationTest {
 
     @Autowired
     private Environment environment;
-
-    @Autowired
-    private List<DatabaseCheckOnHost<? extends DbObject>> checks;
 
     @Test
     void contextLoadsAndDoesNotContainPgIndexHealthBeans() {
@@ -51,17 +43,5 @@ class PostgresDemoApplicationTest {
             .isNotBlank()
             .startsWith("jdbc:postgresql://localhost:")
             .endsWith("/demo_for_pg_index_health_starter?loggerLevel=OFF");
-    }
-
-    @Test
-    void checksShouldWork() {
-        assertThat(checks)
-            .hasSameSizeAs(Diagnostic.values());
-
-        checks.stream()
-            .filter(DatabaseCheckOnHost::isStatic)
-            .forEach(c -> assertThat(c.check())
-                .as(c.getDiagnostic().name())
-                .isEmpty());
     }
 }
