+++ spring-boot-integration/postgres-tc-url-demo-app/src/test/java/io/github/mfvanek/pg/spring/postgres/tc/url/PostgresTestcontainersUrlDemoApplicationTest.java
@@ -12,9 +12,6 @@ package io.github.mfvanek.pg.spring.postgres.tc.url;
 
 import com.zaxxer.hikari.HikariDataSource;
 import io.github.mfvanek.pg.connection.PgConnection;
-import io.github.mfvanek.pg.core.checks.common.DatabaseCheckOnHost;
-import io.github.mfvanek.pg.core.checks.common.Diagnostic;
-import io.github.mfvanek.pg.model.dbobject.DbObject;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -22,12 +19,10 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.core.env.Environment;
 import org.springframework.test.context.ActiveProfiles;
 
-import java.util.List;
-
 import static org.assertj.core.api.Assertions.assertThat;
 
-@ActiveProfiles("test")
 @SpringBootTest
+@ActiveProfiles("test")
 class PostgresTestcontainersUrlDemoApplicationTest {
 
     @Autowired
@@ -35,9 +30,6 @@ class PostgresTestcontainersUrlDemoApplicationTest {
 
     @Autowired
     private Environment environment;
-
-    @Autowired
-    private List<DatabaseCheckOnHost<? extends DbObject>> checks;
 
     @Test
     void contextLoadsAndDoesNotContainPgIndexHealthBeans() {
@@ -50,23 +42,5 @@ class PostgresTestcontainersUrlDemoApplicationTest {
         assertThat(environment.getProperty("spring.datasource.url"))
             .isNotBlank()
             .isEqualTo("jdbc:tc:postgresql:17.2:///demo_for_pg_index_health_starter");
-    }
-
-    @Test
-    void checksShouldWork() {
-        assertThat(checks)
-            .hasSameSizeAs(Diagnostic.values());
-
-        checks.stream()
-            .filter(DatabaseCheckOnHost::isStatic)
-            .forEach(c -> {
-                assertThat(c.check())
-                    .as(c.getDiagnostic().name())
-                    .isEmpty();
-
-                assertThat(c.getHost().getPgUrl())
-                    .startsWith("jdbc:postgresql://localhost:")
-                    .endsWith("/test?loggerLevel=OFF");
-            });
     }
 }
